New Inlined Super High Performance MD5 Algorithm
This time I had the idea to pass the JavaScript code through a C
PreProcessor so I can use C-Style Macros to easily inline code. Turns
out it was worth it with almost a doubling in MD5 performance! On the
downside it does create less readable OUTPUT code but the source
preprocessor code is just as easy to read and understand!